backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
python-socketio
google-genai
google-cloud-storage
gTTS
python-dotenv
pytest
pytest-asyncio
numpy
